Update Report-Msgid-Bugs-To field in l10n files
Previously the Report-Msgid-Bugs-To Header Entry field was empty. I
added a link to the project's issue tracker on GitHub in all PO files
so that people can know where to go to report problems with the
untranslated strings. I also added the link to the POT file so that new
translation files created from it will have the link as well.