Update Project-Id-Version field in l10n files
Previously the Project-Id-Version field was set to "PACKAGE VERSION",
which caused msgfmt to give the warning "header field
'Project-Id-Version' still has the initial default value", so I set the
project name and version in the POT file and updated the PO files
accordingly. The German PO file actually already had its own text but I
updated it to match the others for consistency.