Prefix all KDE4 app names with kde4-
KDE apps' "app" names are prefixed with kde4, e.g. kde4-kmess instead of
kmess. Without this they will not be listed in the messaging menu unless
you symlink the .desktop file from /usr/share/applications/kde4 to
/usr/share/applications.

I've prefixed the KDE4 apps with kde4 now, installed the apps and
verified that they all appear in the menu after this fix and do not
appear without it. I tested this in Gnome Shell 3.12 on ArchLinux. If
this breaks listing of KDE apps on other distributions or shell versions
then we should duplicate each KDE app, once with the prefix and once
without.

This fixes #15.