Stop tracking locale binaries, build them with `make` instead
It's not so nice to track binaries with git, they can't be diffed and
they shouldn't be edited because they're not source code. These files
also easily got out of date with their source.

I've deleted the locale binaries (.mo files under locale/) and told git
not to track sub-directories of locale/. People who want to generate the
locales for testing should generate them using the new 'locales' make
target from the top top directory, i.e.:

    `make locales`

I've left a note about this in a README file in locale/ which
conveniently keeps the directory itself in git so that people who want
to generate locales can know where to look.