Stop tracking gschemas.compiled, build it with make instead
Just like the compiled locale binaries, I've removed the
gschemas.compiled file and told git to ignore it. I've added a new make
target 'schemas' which compiles the gschemas.compiled file from files
found in the schemas/ directory, and also an 'all' target which compiles
both the schemas and the locale binaries.