Install extension to UUID-named sub-directory
This is a fix for the `install` make target.  Installing this extension
through the make target would unpack all the files to the install
directory root, meaning it wouldn't be found and the directory would be
littered with extension files.

This commit fixes this by adding the UUID as the last part of the target
directory for unzipping, after the general installation directory for
extensions.  Now instead of:

    .local/share/gnome-shell/extensions

the extension files are unpacked to:

    .local/share/gnome-shell/extensions/messagingmenu@screenfreeze.net

Tested on my own laptop, seems to work fine, can't believe this has been
a bug for so long and I didn't notice.